rename: 폴더명 Form -> Register로 변경된 path 적용
