+++ src/components/Register/InputGroup/index.tsx
@@ -5,8 +5,8 @@ import { emailOptions } from 'src/utils/emailListUtil';
 import { IAuthForm } from 'src/types/registerPage';
 import { useDaumPostcodePopup } from 'react-daum-postcode';
 import { postcodeScriptUrl } from 'react-daum-postcode/lib/loadPostcode';
-import Styled from '@components/Form/styles';
 import FormButton from '../FormButton';
+import Styled from '../styles';
 
 const InputGroup = () => {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
