+++ src/components/Register/SnsCircles/index.tsx
