+++ src/components/Register/SnsCircles/styles.tsx
